formatting

+++ 22330_Powerplay_Android_Studio/TeamCode/src/main/java/org/firstinspires/ftc/teamcode/teleop/ManualOrientation.java
@@ -1,13 +1,9 @@
 package org.firstinspires.ftc.teamcode.teleop;
 
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
-import com.qualcomm.robotcore.hardware.TouchSensor;
-import com.qualcomm.hardware.rev.RevHubOrientationOnRobot;
-import com.qualcomm.robotcore.hardware.IMU;
-import com.qualcomm.robotcore.hardware.CRServo;
-import com.qualcomm.robotcore.hardware.Servo;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 import com.qualcomm.robotcore.hardware.DcMotor;
+import com.qualcomm.robotcore.hardware.Servo;
 
 @TeleOp(name = "Manual Orientation", group = "Experimental")
 public class ManualOrientation extends LinearOpMode {
@@ -21,12 +17,12 @@ public class ManualOrientation extends LinearOpMode {
     private double frontRightPower;
     private double backLeftPower;
     private double backRightPower;
-    
+
     private double fLX = 1;
     private double fRX = 1;
     private double bLX = 1;
     private double bRX = 1;
-    
+
     private double fLY = 1;
     private double fRY = 1;
     private double bLY = 1;
@@ -41,7 +37,7 @@ public class ManualOrientation extends LinearOpMode {
     // Lift systems
     private DcMotor lift;
     private boolean liftWasStatic = false;
-    
+
     // Claw systems
     private Servo claw;
 
@@ -51,22 +47,22 @@ public class ManualOrientation extends LinearOpMode {
             normal = 1;
         }
         normal /= percent;
-        
-        frontLeft.setPower(((fLY*y)+(fLX*x)+pivot)/normal);
-        frontRight.setPower(((fRY*y)-(fRX*x)-pivot)/normal);
-        backLeft.setPower(((bLY*y)-(bLX*x)+pivot)/normal);
-        backRight.setPower(((bRY*y)+(bRX*x)-pivot)/normal);
+
+        frontLeft.setPower(((fLY * y) + (fLX * x) + pivot) / normal);
+        frontRight.setPower(((fRY * y) - (fRX * x) - pivot) / normal);
+        backLeft.setPower(((bLY * y) - (bLX * x) + pivot) / normal);
+        backRight.setPower(((bRY * y) + (bRX * x) - pivot) / normal);
     }
 
     @Override
     public void runOpMode() {
-        frontLeft = (DcMotor)hardwareMap.get("frontLeft");
-        frontRight = (DcMotor)hardwareMap.get("frontRight");
-        backLeft = (DcMotor)hardwareMap.get("backLeft");
-        backRight = (DcMotor)hardwareMap.get("backRight");
+        frontLeft = (DcMotor) hardwareMap.get("frontLeft");
+        frontRight = (DcMotor) hardwareMap.get("frontRight");
+        backLeft = (DcMotor) hardwareMap.get("backLeft");
+        backRight = (DcMotor) hardwareMap.get("backRight");
 
-        lift = (DcMotor)hardwareMap.get("lift");
-        claw = (Servo)hardwareMap.get("claw");
+        lift = (DcMotor) hardwareMap.get("lift");
+        claw = (Servo) hardwareMap.get("claw");
 
         frontLeft.setDirection(DcMotor.Direction.FORWARD);
         frontRight.setDirection(DcMotor.Direction.REVERSE);
@@ -97,12 +93,12 @@ public class ManualOrientation extends LinearOpMode {
                 lift.setPower(-gamepad1.right_stick_y);
             } else if (!liftWasStatic) {
                 lift.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
-                lift.setTargetPosition((int) 0);
+                lift.setTargetPosition(0);
                 lift.setMode(DcMotor.RunMode.RUN_TO_POSITION);
                 lift.setPower(0.3);
                 liftWasStatic = true;
             }
-                    
+
             if (gamepad1.left_bumper) {
                 if (!gamepad1.right_bumper) {
                     claw.setPosition(0);
@@ -110,20 +106,20 @@ public class ManualOrientation extends LinearOpMode {
             } else if (gamepad1.right_bumper) {
                 claw.setPosition(1);
             }
-            
+
             if (gamepad1.y) {
                 percent = 1;
             }
             if (gamepad1.a) {
                 percent = 0.9;
             }
-            
+
             if (gamepad1.y) {
                 fLX = 1;
                 fRX = 1;
                 bLX = 1;
                 bRX = 1;
-                
+
                 fLY = 1;
                 fRY = 1;
                 bLY = 1;
@@ -134,7 +130,7 @@ public class ManualOrientation extends LinearOpMode {
                 fRX = -1;
                 bLX = -1;
                 bRX = -1;
-                
+
                 fLY = -1;
                 fRY = -1;
                 bLY = -1;
@@ -145,7 +141,7 @@ public class ManualOrientation extends LinearOpMode {
                 fRX = 1;
                 bLX = 1;
                 bRX = -1;
-                
+
                 fLY = 1;
                 fRY = -1;
                 bLY = -1;
@@ -156,7 +152,7 @@ public class ManualOrientation extends LinearOpMode {
                 fRX = -1;
                 bLX = -1;
                 bRX = 1;
-                
+
                 fLY = -1;
                 fRY = 1;
                 bLY = 1;
