+++ 22330_Powerplay_Android_Studio/TeamCode/src/main/java/org/firstinspires/ftc/teamcode/teleop/TestOp.java
@@ -1,9 +1,8 @@
 package org.firstinspires.ftc.teamcode.teleop;
 
+import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 import com.qualcomm.robotcore.hardware.DcMotor;
-import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
-import com.qualcomm.robotcore.hardware.HardwareMap;
 
 @TeleOp(name = "TestOp", group = "Experimental")
 public class TestOp extends LinearOpMode {
@@ -11,11 +10,11 @@ public class TestOp extends LinearOpMode {
 
     private DcMotor arm;
     private boolean armWasStatic = false;
-    
+
     @Override
     public void runOpMode() {
-        shoulder = (DcMotor)hardwareMap.get("shoulder");
-        arm = (DcMotor)hardwareMap.get("arm");
+        shoulder = (DcMotor) hardwareMap.get("shoulder");
+        arm = (DcMotor) hardwareMap.get("arm");
 
         arm.setDirection(DcMotor.Direction.REVERSE);
 
@@ -24,18 +23,18 @@ public class TestOp extends LinearOpMode {
             // Wat duh hek
             // Y values are negative when stick points up
             // positive when stick points down?!?!?!
-            
-            shoulder.setPower(gamepad1.right_stick_y/2);
+
+            shoulder.setPower(gamepad1.right_stick_y / 2);
 
             if (-gamepad1.right_stick_y != 0) {
                 if (armWasStatic) {
                     armWasStatic = false;
                     arm.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
                 }
-                arm.setPower(-gamepad1.left_stick_y/2);
+                arm.setPower(-gamepad1.left_stick_y / 2);
             } else if (!armWasStatic) {
                 arm.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
-                arm.setTargetPosition((int) 0);
+                arm.setTargetPosition(0);
                 arm.setMode(DcMotor.RunMode.RUN_TO_POSITION);
                 arm.setPower(0.3);
                 armWasStatic = true;
